Updated calculation of restriction string and adding of bands
Before it was required that each WAP had a "Max daily rate pro rata"
field. This has changed since the DB update with diverts being conversed
to takes if the water leaves the river bed. The diverts therefore do not
have the "Max daily rate pro rata" field anymore. This required an
update of the "Restriction daily volume" branch.

The other update is related to the addding of bands. In the previous
version bands were always integers. With the rangita model this can be a
string. The int() function has therefore been removed.